requirements.txt
flask==2.2.2
requests==2.28.1
selenium==4.4.3
bs4==0.0.1
pytube==12.1.0
pymongo
mysql-connector-python